Fix clippy pedantic warnings

+++ src/day09.rs
@@ -1,9 +1,9 @@
 use std::collections::{HashMap, HashSet, VecDeque};
 
-type Grid = HashMap<(i32, i32), u32>;
-type Point = (i32, i32);
+type Grid = HashMap<(i64, i64), u32>;
+type Point = (i64, i64);
 
-fn get_neighbors(p: Point, y_max: i32, x_max: i32) -> Vec<Point> {
+fn get_neighbors(p: Point, y_max: i64, x_max: i64) -> Vec<Point> {
     [(-1, 0), (1, 0), (0, -1), (0, 1)]
         .iter()
         .map(|&(dy, dx)| (p.0 + dy, p.1 + dx))
@@ -11,7 +11,7 @@ fn get_neighbors(p: Point, y_max: i32, x_max: i32) -> Vec<Point> {
         .collect()
 }
 
-fn get_lows(map: &Grid, y_max: i32, x_max: i32) -> Vec<Point> {
+fn get_lows(map: &Grid, y_max: i64, x_max: i64) -> Vec<Point> {
     let mut low_points = Vec::new();
     itertools::iproduct!(0..y_max, 0..x_max).for_each(|p| {
         if get_neighbors(p, y_max, x_max)
@@ -24,7 +24,7 @@ fn get_lows(map: &Grid, y_max: i32, x_max: i32) -> Vec<Point> {
     low_points
 }
 
-fn get_basin_size(map: &Grid, y_max: i32, x_max: i32, start: Point) -> i64 {
+fn get_basin_size(map: &Grid, y_max: i64, x_max: i64, start: Point) -> i64 {
     let mut visited = HashSet::new();
     let mut queue = VecDeque::new();
     queue.push_back(start);
@@ -52,10 +52,10 @@ pub fn solve_2() -> i64 {
 
     for (y, line) in include_str!("../input/09.txt").lines().enumerate() {
         for (x, c) in line.chars().map(|c| c.to_digit(10).unwrap()).enumerate() {
-            map.entry((y as i32, x as i32)).or_insert(c);
+            map.entry((y as i64, x as i64)).or_insert(c);
         }
         y_max += 1;
-        x_max = line.chars().count() as i32;
+        x_max = line.chars().count() as i64;
     }
 
     let mut basins = get_lows(&map, y_max, x_max)
@@ -74,14 +74,14 @@ pub fn solve_1() -> i64 {
 
     for (y, line) in include_str!("../input/09.txt").lines().enumerate() {
         for (x, c) in line.chars().map(|c| c.to_digit(10).unwrap()).enumerate() {
-            map.entry((y as i32, x as i32)).or_insert(c);
+            map.entry((y as i64, x as i64)).or_insert(c);
         }
         y_max += 1;
-        x_max = line.chars().count() as i32;
+        x_max = line.chars().count() as i64;
     }
 
     get_lows(&map, y_max, x_max)
         .iter()
-        .map(|p| (map.get(p).unwrap() + 1) as i64)
+        .map(|p| i64::from(map.get(p).unwrap() + 1))
         .sum()
 }
